Barra de progresso melhorada
Agora é possivel escolher a posição inicial.
Bug no metodo getDuracaoTique corrigido: a divisao segundos/tiques
ocorria antes da conversao para double, resultando em 0.
Alteração no metodo mudarPosicao: o controle de quando tocar deve ficar
a cargo da GUI.

+++ src/Trabalho1/GestorArquivo.java
@@ -62,7 +62,7 @@ public class GestorArquivo {
 		long microsegundos = sequencia.getMicrosecondLength();
 		long segundos = microsegundos/1000000;
 		long  tiques    = sequencia.getTickLength();
-		return segundos/tiques;
+		return (double)segundos/(double)tiques;
 	}
 	
 	public Receiver getReceptor(){
